+++ src/api/ffi/symbols/scheduler.ts
@@ -1,21 +1,21 @@
 export const schedulerSymbols = {
   linear_decay_scheduler: {
     parameters: [
-      "f64",
+      "f32",
       "usize",
     ],
     result: "pointer",
   } as const,
   exponential_decay_scheduler: {
     parameters: [
-      "f64",
+      "f32",
       "usize",
     ],
     result: "pointer",
   } as const,
   one_cycle_scheduler: {
     parameters: [
-      "f64",
+      "f32",
       "usize",
     ],
     result: "pointer",
